Align on a single startup sequence
