+++ gotham/src/helpers/http/header/x_content_type_options.rs
@@ -2,7 +2,7 @@
 
 use std::fmt;
 // TODO: Remove when this import isn't required in stable anymore.
-#[allow(unused_imports)]
+#[allow(deprecated, unused_imports)]
 use std::ascii::AsciiExt;
 
 use hyper;
