+++ gotham/src/lib.rs
@@ -48,25 +48,71 @@ pub mod router;
 mod service;
 pub mod state;
 pub mod test;
-mod os;
-
-pub use os::current::start_with_num_threads;
 
 use std::net::{SocketAddr, TcpListener, ToSocketAddrs};
+use std::sync::Arc;
+use std::thread;
+
+use futures::{Future, Stream};
+use hyper::server::Http;
+use tokio_core::reactor::Core;
+
 use handler::NewHandler;
+use service::GothamService;
 
 /// Starts a Gotham application, with the default number of threads (equal to the number of CPUs).
-///
-/// ## Windows
-///
-/// An additional thread is used on Windows to accept connections.
 pub fn start<NH, A>(addr: A, new_handler: NH)
 where
     NH: NewHandler + 'static,
     A: ToSocketAddrs,
 {
     let threads = num_cpus::get();
-    start_with_num_threads(addr, threads, new_handler)
+    let (listener, addr) = ::tcp_listener(addr);
+
+    let protocol = Arc::new(Http::new());
+    let new_handler = Arc::new(new_handler);
+
+    info!(
+        target: "gotham::start",
+        " Gotham listening on http://{} with {} threads",
+        addr,
+        threads,
+    );
+
+    for thread_n in 0..threads - 1 {
+        let listener = listener.try_clone().expect("unable to clone TCP listener");
+        let protocol = protocol.clone();
+        let new_handler = new_handler.clone();
+        thread::Builder::new()
+            .name(format!("gotham-{}", thread_n))
+            .spawn(move || start_core(listener, &addr, &protocol, new_handler))
+            .expect("unable to spawn thread");
+    }
+
+    start_core(listener, &addr, &protocol, new_handler);
+}
+
+fn start_core<NH>(listener: TcpListener, addr: &SocketAddr, protocol: &Http, new_handler: Arc<NH>)
+where
+    NH: NewHandler + 'static,
+{
+    let mut core = Core::new().expect("unable to spawn tokio reactor");
+    let handle = core.handle();
+
+    let gotham_service = GothamService::new(new_handler, handle.clone());
+
+    let listener = tokio_core::net::TcpListener::from_listener(listener, addr, &handle)
+        .expect("unable to convert TCP listener to tokio listener");
+
+    let f = Box::new(listener.incoming().for_each(move |(socket, addr)| {
+        let service = gotham_service.connect(addr);
+        let f = protocol.serve_connection(socket, service).then(|_| Ok(()));
+
+        handle.spawn(f);
+        Ok(())
+    }));
+
+    core.run(f).expect("unable to run reactor over listener");
 }
 
 fn tcp_listener<A>(addr: A) -> (TcpListener, SocketAddr)
