--- gotham/src/os/mod.rs
@@ -1,9 +1,0 @@
-#[cfg(not(windows))]
-pub mod unix;
-#[cfg(not(windows))]
-pub use self::unix as current;
-
-#[cfg(windows)]
-pub mod windows;
-#[cfg(windows)]
-pub use self::windows as current;
